requirements.txt
# Procesamiento de texto y PDF
PyPDF2>=3.0.0

# Procesamiento de imágenes (opcional, solo si procesarás imágenes)
pillow>=9.0.0
pytesseract>=0.3.10

# Machine learning (versión ligera de PyTorch)
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cpu

# Procesamiento de texto
sentence-transformers>=2.2.2
numpy>=1.21
scikit-learn>=1.0

# Google AI
google-generativeai>=0.3.0

# Interfaz web
streamlit>=1.20.0

# Manejo de variables de entorno
python-dotenv>=1.0.0